css-align-3_dev/xhtml1print/place-items-shorthand-004.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Box Alignment: place-items shorthand - invalid values</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez" />
<link href="http://www.w3.org/TR/css3-align/#place-items-property" rel="help" />
<meta content="Check that place-items's invalid values are properly detected at parsing time." name="assert" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-items: start;
        justify-items: end;
    }
</style>
</head><body><div id="log"></div>
<div id="test"></div>
<script>
    function checkInvalidValues(value) {
        var div = document.getElementById("test");
        div.setAttribute("style", "place-items: " + value);
        var style = getComputedStyle(div);
        assert_equals(style.getPropertyValue("align-items"),
            "start", "align-items computed value");
        assert_equals(style.getPropertyValue("justify-items"),
            "end", "justify-items computed value");
    }

    test(function() {
        checkInvalidValues("center safe")
        checkInvalidValues("true center")
    }, "Verify overflow keywords are invalid");

    test(function() {
        checkInvalidValues("center space-between start")
    }, "Verify fallback values are invalid");

    test(function() {
        checkInvalidValues("10px left")
        checkInvalidValues("right 10%")
    }, "Verify numeric values are invalid");

    test(function() {
       checkInvalidValues("auto right")
       checkInvalidValues("auto auto")
    }, "Verify 'auto' values are invalid");

    test(function() {
        checkInvalidValues("")
    }, "Verify empty declaration is invalid");
</script>
</body></html>
